Moved ReplacementTypes to util from lang and replaced spaces with tabs in the tests (and updated the switch statement)
