--- src/main/java/ch/njol/skript/lang/ReplacementTypes.java
@@ -1,7 +1,0 @@
-package ch.njol.skript.lang;
-
-public enum ReplacementTypes {
-	ALL,
-	FIRST,
-	LAST
-}
