+++ src/test/skript/tests/syntaxes/expressions/ExprReplace.sk
@@ -21,10 +21,10 @@ test "replace":
 	set {_text::9} to case-sensitive replace last "BOB" with "bobby" in {_text::9}
 	assert {_text::1} is "My name is bobby and this is my friend bobby! Say hi, bobby!" with "Replace all failed"
 	assert {_text::2} is "My name is bobby and this is my friend bob! Say hi, bob!" with "Replace first failed"
-    assert {_text::3} is "My name is bob and this is my friend bob! Say hi, bobby!" with "Replace last failed"
-    assert {_text::4} is "My name is BOBBY and this is my friend BOBBY! Say hi, BOBBY!" with "1st Replace all case-sensitive failed"
-    assert {_text::5} is "My name is bob and this is my friend bob! Say hi, bob!" with "2nd Replace all case-sensitive failed"
-    assert {_text::6} is "My name is BOBBY and this is my friend bob! Say hi, bob!" with "1st Replace first case-sensitive failed"
-    assert {_text::7} is "My name is bob and this is my friend bob! Say hi, bob!" with "2nd Replace first case-sensitive failed"
-    assert {_text::8} is "My name is bob and this is my friend bob! Say hi, BOBBY!" with "1st Replace last case-sensitive failed"
-    assert {_text::9} is "My name is bob and this is my friend bob! Say hi, bob!" with "2nd Replace last case-sensitive failed"
+	assert {_text::3} is "My name is bob and this is my friend bob! Say hi, bobby!" with "Replace last failed"
+	assert {_text::4} is "My name is BOBBY and this is my friend BOBBY! Say hi, BOBBY!" with "1st Replace all case-sensitive failed"
+	assert {_text::5} is "My name is bob and this is my friend bob! Say hi, bob!" with "2nd Replace all case-sensitive failed"
+	assert {_text::6} is "My name is BOBBY and this is my friend bob! Say hi, bob!" with "1st Replace first case-sensitive failed"
+	assert {_text::7} is "My name is bob and this is my friend bob! Say hi, bob!" with "2nd Replace first case-sensitive failed"
+	assert {_text::8} is "My name is bob and this is my friend bob! Say hi, BOBBY!" with "1st Replace last case-sensitive failed"
+	assert {_text::9} is "My name is bob and this is my friend bob! Say hi, bob!" with "2nd Replace last case-sensitive failed"
